Straightforward generalization for alternate text representations
This patch adds a Chars class, which has several operations from the
ListLike class.  It updates all the code so that instead of the String
type any instance of Chars can be used.  A Text.PrettyPrint.String
module is added which specializes the public API for the String type.
This module is re-exported by Text.PrettyPrint.

In order to avoid new dependencies, no instance for Text is included
in library, but the instance is provided in the tests directory.  For
the same reason, the ListLike class is not used.

It may be that older versions of GHC will not accept the new
extensions required to compile these changes.